update pixel firing location to be globally available
